node-main/deps/v8/test/inspector/heap-profiler/heap-snapshot-js-weak-refs-expected.txt

Tests weakness of edges from JSWeakRef and WeakCell.

Running test: testHeapSnapshotJSWeakRefs
WeakRef target edge type: weak
WeakCell target edge type: weak
WeakCell holdings edge type: hidden
WeakCell unregister token edge type: weak
